with dao (but dao not completed)

+++ product/clothes/types/Pants.java
@@ -4,8 +4,7 @@ import com.flipkart.product.clothes.Clothes;
 
 public class Pants extends Clothes {
 
-    public Pants(final String gender, final String size, final int price, final String brand) {
+    public Pants(final String gender, final String size, final float price, final String brand) {
         super("Pants", gender, size, price, brand);
     }
-
 }
