+++ product/clothes/types/Shirt.java
@@ -4,7 +4,7 @@ import com.flipkart.product.clothes.Clothes;
 
 public class Shirt extends Clothes {
 
-    public Shirt(final String gender, final String size, final int price, final String brand) {
+    public Shirt(final String gender, final String size, final float price, final String brand) {
         super("Tshirt", gender, size, price, brand);
     }
 }
