+++ product/laptop/brands/AcerLaptop.java
@@ -3,7 +3,7 @@ package com.flipkart.product.laptop.brands;
 import com.flipkart.product.laptop.Laptop;
 
 public class AcerLaptop extends Laptop {
-    public AcerLaptop(final String model, final int price) {
+    public AcerLaptop(final String model, final float price) {
         super("Acer", model, price);
     }
 }
