+++ product/laptop/brands/DellLaptop.java
@@ -3,7 +3,7 @@ package com.flipkart.product.laptop.brands;
 import com.flipkart.product.laptop.Laptop;
 
 public class DellLaptop extends Laptop {
-    public DellLaptop(final String model, final int price) {
+    public DellLaptop(final String model, final float price) {
         super("Samsung", model, price);
     }
 }
