+++ product/laptop/brands/HpLaptop.java
@@ -3,7 +3,7 @@ package com.flipkart.product.laptop.brands;
 import com.flipkart.product.laptop.Laptop;
 
 public class HpLaptop extends Laptop {
-    public HpLaptop(final String model, final int price) {
+    public HpLaptop(final String model, final float price) {
         super("Hp", model, price);
     }
 }
