+++ product/mobile/brands/SamsungMobile.java
@@ -4,7 +4,7 @@ import com.flipkart.product.mobile.Mobile;
 
 public class SamsungMobile extends Mobile {
 
-    public SamsungMobile(final String model, final int price) {
+    public SamsungMobile(final String model, final float price) {
         super("Samsung", model, price);
     }
 }
